requirements.txt
matplotlib==3.9.4
Pillow==11.1.0
PyYAML==6.0.2
gdown==5.2.0
dvc==3.59.0
dvc-gdrive==3.0.1
mlflow==2.20.2
Flask==3.1.0
prometheus-client==0.19.0
prometheus-flask-exporter==0.23.0
psutil==5.9.5
pandas==2.0.3
scikit-learn==1.3.0
joblib==1.3.2
numpy==1.26.4